requirements.txt
Flask==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
# Use a version compatible with Python 3.13
scikit-learn>=1.6.0
numpy>=2.0.0
pandas>=2.2.2
# Optional
# torch==2.3.1
waitress>=3.0.0
# For Gemini chatbot
google-generativeai>=0.3.0
requests>=2.31.0
